refactor(conversation): extract shared FilePreviewContent component (#28023)
